// generated/src/app/api-models/codereview-result.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class CodereviewResult {
  seqClaimResultCode: number;
  claimResultCode: string;
  moduleType: string;
  crStatus: string;
  recNum: number;
  claimAction: string;
  reasnCode: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}